requirements.txt
psycopg2
requests
msal
python-dotenv
flask
